This changes maps from bytes to ushorts to allow more than 255 blocks, to load maps that still go by byte standard, you now have to use /lb

+++ Commands/CmdAbout.cs
@@ -48,10 +48,10 @@ namespace MCForge.Commands
             Player.SendMessage(p, "/about - Displays information about a block.");
         }
 
-        public void AboutBlockchange(Player p, ushort x, ushort y, ushort z, byte type)
+        public void AboutBlockchange(Player p, ushort x, ushort y, ushort z, ushort type)
         {
             if (!p.staticCommands) p.ClearBlockchange();
-            byte b = p.level.GetTile(x, y, z);
+            ushort b = p.level.GetTile(x, y, z);
             if (b == Block.Zero) { Player.SendMessage(p, "Invalid Block(" + x + "," + y + "," + z + ")!"); return; }
             p.SendBlockchange(x, y, z, b);
