+++ Commands/CmdAscend.cs
@@ -39,7 +39,7 @@ namespace MCForge.Commands
             while (!found && posy < max)
             {
                 posy++;
-                byte block = p.level.GetTile(xpos, posy, zpos);
+                ushort block = p.level.GetTile(xpos, posy, zpos);
                 if (block == Block.air || block == Block.air_door || block == Block.air_switch || block == Block.Zero)
                 {
                     ushort blockabove = (ushort) (posy + 1);
