+++ Commands/CmdMissile.cs
@@ -155,14 +155,14 @@ namespace MCForge.Commands
 			}));
 			aimThread.Start();
 		}
-		public void Blockchange1(Player p, ushort x, ushort y, ushort z, byte type)
+		public void Blockchange1(Player p, ushort x, ushort y, ushort z, ushort type)
 		{
 			if (!p.staticCommands)
 			{
 				p.ClearBlockchange();
 				p.aiming = false;
 			}
-			byte by = p.level.GetTile(x, y, z);
+			ushort by = p.level.GetTile(x, y, z);
 			p.SendBlockchange(x, y, z, by);
 			Pos bp = (Pos)p.blockchangeObject;
